refactor: move body to dataset
