+++ pims/pims/api/utils/deps.py
@@ -1,0 +1,6 @@
+from pims.importer.importer import DatasetImporter
+
+
+def get_dataset_importer() -> DatasetImporter:
+    return DatasetImporter()
+
