+++ pims/pims/importer/utils.py
@@ -1,8 +1,22 @@
 import os
 
+from pims.files.file import Path
+
 REQUIRED_DIRECTORIES = {
     "IMAGES",
 }
+
+
+def is_already_imported(image_path: Path, data_path: Path) -> bool:
+    """Check if an image was already imported."""
+
+    for upload_dir in data_path.iterdir():
+        if not upload_dir.is_dir():
+            continue
+        for candidate in upload_dir.iterdir():
+            if candidate.is_symlink() and candidate.resolve() == image_path.resolve():
+                return True
+    return False
 
 
 def check_dataset_structure(dataset_root: str) -> tuple[bool, list[str]]:
